requirements.txt
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiohttp-jinja2==1.5
aiohttp-session==2.12.0
aiokafka==0.7.2
aiosignal==1.2.0
alabaster==0.7.12
aniso8601==9.0.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
async-timeout==4.0.2
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
bleach==4.1.0
blinker==1.4
brial==1.0.2
Brlapi==0.8.3
Brotli==1.0.9
capstone==5.0.0rc2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
colored-traceback==0.3.0
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
crayons==0.4.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cypari2==2.1.2
Cython==0.29.28
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0
fasteners==0.14.1
feedgenerator==2.0.0
ffmpeg-python==0.2.0
filelock==3.6.0
Flask==2.2.2
Flask-WTF==1.0.1
fonttools==4.29.1
fpylll==0.5.2.dev0
frozenlist==1.3.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gevent==22.10.1
gitdb==4.0.9
GitPython==3.1.29
gmpy2==2.1.2
graphene==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==1.1.3.post0
grequests==0.6.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.22.2
imageio-ffmpeg==0.4.7
imagesize==1.3.0
importlib-metadata==4.6.4
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kafka-python==2.0.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
memory-allocator==0.1.2
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpmath==0.0.0
multidict==6.0.2
nautilus==0.5.2
nautilus-open-any-terminal==0.4.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.4
nose2==0.12.0
notebook==6.4.8
number-parser==0.3.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pathlib2==2.3.7.post1
peewee==3.15.3
pelican==4.8.0
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.0.1
pipreqs==0.4.11
pkgconfig==1.5.5
platformdirs==2.5.1
pluggy==0.13.0
plumbum==1.7.2
ply==3.11
pplpy==0.8.7
primecountpy==0.1.0
proglog==0.1.10
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
pwntools==4.8.0
py==1.10.0
pycairo==1.20.1
pycryptodome==3.15.0
pycups==2.0.1
pydantic==1.10.2
pyelftools==0.29
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.0
python-apt==2.3.0+ubuntu2.1
python-consul==1.1.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
rich==12.6.0
roman==3.3
ROPGadget==7.1
rpy2==3.4.5
rpyc==5.2.3
sagemath-standard==9.5
sagenb-export==3.2
scapy==2.4.5
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
Sphinx==4.3.2
stegano==0.10.2
stepic==0.5.0
sympy==1.9
systemd-python==234
tabulate==0.9.0
terminado==0.13.1
terminator==2.1.1
testpath==0.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tox==3.21.4
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.4.0
tzlocal==2.1
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unicorn==2.0.0
Unidecode==1.3.6
urllib3==1.26.5
usb-creator==0.3.7
uvloop==0.17.0
vboxapi==1.0
vec==0.5
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==2.0.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.8.1
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.5.0
